--- integration-test/src/thriftserver/main/scala/org/apache/livy/test/framework/ThriftJdbcClient.scala
@@ -1,51 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.livy.test.framework
-
-import java.sql.{Connection, DriverManager, ResultSet}
-
-class ThriftJdbcClient(val jdbcUri: String) {
-
-  private def newConnection: Connection = {
-    val user = System.getProperty("user.name")
-    DriverManager.getConnection(jdbcUri, user, "")
-  }
-
-  def checkQuery(connection: Connection, query: String)(validate: ResultSet => Unit): Unit = {
-    val ps = connection.prepareStatement(query)
-    try {
-      val rs = ps.executeQuery()
-      try {
-        validate(rs)
-      } finally {
-        rs.close()
-      }
-    } finally {
-      ps.close()
-    }
-  }
-
-  def withConnection[T](f: Connection => T): T = {
-    val connection = newConnection
-    try {
-      f(connection)
-    } finally {
-      connection.close()
-    }
-  }
-}
